requirements.txt
annotated-types==0.7.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.1
cymem==2.0.11
defusedxml==0.7.1
filelock==3.17.0
fonttools==4.56.0
fpdf2==2.8.2
fsspec==2025.2.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
libmagic==1.0
llvmlite==0.44.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pluggy==1.5.0
preshed==3.0.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pypdfium2==4.30.1
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-magic==0.4.27
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
setuptools==75.8.0
shellingham==1.5.4
shiboken6==6.8.2.1
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.3
thinc==8.3.4
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
